requirements.txt
Flask
Pylette
Pillow
